perf: Trade range repartition + symbols cast + audit table + threshold bump
- trade.py: partition the spark.range() base for trade_df to
  max(8, cfg.sf // 8) so each task handles ~1M rows. Default partitioning
  was putting ~108M rows/task at SF=20000, spilling 237+ GB inside the
  "range" pipeline as 30+ withColumns and the _symbols join piled up.

- finwire.py: cast _symbols._idx to long before caching. Downstream
  consumers compare it to hash_key() expressions (bigint), and the
  implicit cast Spark inserts can prevent auto-broadcast even when
  _symbols fits under the threshold.

- spark_data_generator.py: raise spark.sql.autoBroadcastJoinThreshold
  and spark.databricks.adaptive.autoBroadcastJoinThreshold from 200m
  to 250m. _symbols' actual shuffle size at SF=20000 is ~200 MB, but
  the planner's per-row size estimate trips the 200MB threshold even
  when runtime stats would broadcast fine.

- dw_init.sql (incremental): Audit is now a materialized TABLE, not a
  VIEW over the *_audit.csv files in the volume. Previously if the next
  datagen run wiped the volume mid-benchmark, the Audit view lost its
  source files and the validation queries failed. Materializing the
  rows into Delta decouples audit validation from file lifecycle.

+++ src/tools/tpcdi_gen/finwire.py
@@ -403,8 +403,14 @@ def generate(spark: SparkSession, cfg, dicts: dict, dbutils, symbols_ready_event
         # extend into symbols created in their quarter window. Prior
         # alphabetical ordering caused ~25% of WH ACTV pairs to fall back
         # to _sym0 at SF=5000+, producing a -24.5% drift vs DIGen.
-        .withColumn("_idx", F.row_number().over(
-            Window.orderBy("creation_quarter", "Symbol")) - 1)
+        # _idx cast to long here: downstream consumers (Trade, WatchHistory,
+        # DailyMarket) compare this to hash_key(...) expressions that return
+        # long. If _idx stayed int, Spark's analyzer inserts a cast(_idx as
+        # bigint) into the join condition — that cast can prevent the
+        # optimizer from auto-broadcasting _symbols even when it's well under
+        # the autoBroadcastJoinThreshold.
+        .withColumn("_idx", (F.row_number().over(
+            Window.orderBy("creation_quarter", "Symbol")) - 1).cast("long"))
         .select("Symbol", "creation_quarter", "deactivation_quarter", "_idx"))
 
     # Stage _symbols to Parquet so Trade/WatchHistory/DailyMarket can read
